modules/py_image: Move descriptor objects to py_image_descriptor.c.
Moves the cascade, keypoints, lbp and keypoints match objects, along
with the cascade and descriptor functions, into a separate file to
reduce the size of py_image.c.
